Bug: Si la suppression d'un compte-rendu échoue, la  source est quand même supprimée Bug: Erreur de majuscule dans le require du phpThumb.config.php
git-svn-id: svn://svn.code.sf.net/p/mediboard/code/trunk@10111 a21d6be2-e512-0410-a060-91f5a2946631

+++ modules/dPcompteRendu/classes/compteRendu.class.php
@@ -471,8 +471,8 @@ class CCompteRendu extends CDocumentItem {
     foreach($files as $_file) {
       $_file->delete();
     }
-    $this->_ref_content->delete();
-    return parent::delete();
+    if ($msg = parent::delete()) { return $msg; }
+    return $this->_ref_content->delete();
   }
 	
   function handleSend() {
